backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-multipart==0.0.6
email-validator==2.1.0
requests==2.31.0
supabase==2.22.1

